Формируем класс смарт
