+++ src/view/smart.js
@@ -1,0 +1,34 @@
+import AbstractView from './abstract';
+
+export default class Smart extends AbstractView {
+  constructor() {
+    super();
+    this._data = {};
+  }
+
+  updateData(update) {
+    if (!update) {
+      return;
+    }
+
+    this._data = { ...this._data, ...update };
+
+    this.updateElement();
+  }
+
+  updateElement() {
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreHandlers();
+  }
+
+  restoreHandlers() {
+    throw new Error('Abstract method not implemented: resetHandlers');
+  }
+}
